requirements.txt
requests==2.22.0
python-miio==0.4.8
